extract calc total stats into a module
